+++ src/ResourceManager/KeyVault/Commands.KeyVault/Commands/RemoveAzureKeyVaultCertificate.cs
@@ -27,7 +27,7 @@ namespace Microsoft.Azure.Commands.KeyVault
         SupportsShouldProcess = true,
         ConfirmImpact = ConfirmImpact.High,
         HelpUri = Constants.KeyVaultHelpUri)]
-    [OutputType(typeof(KeyVaultCertificate))]
+    [OutputType(typeof(DeletedCertificateBundle))]
     public class RemoveAzureKeyVaultCertificate : KeyVaultCmdletBase
     {
         #region Input Parameter Definitions
@@ -58,6 +58,13 @@ namespace Microsoft.Azure.Commands.KeyVault
         [Parameter(HelpMessage = "Do not ask for confirmation.")]
         public SwitchParameter Force { get; set; }
 
+        /// <summary>
+        /// If present, operate on the deleted key entity.
+        /// </summary>
+        [Parameter( Mandatory = false,
+                    HelpMessage = "Permanently remove the previously deleted certificate." )]
+        public SwitchParameter InRemovedState { get; set; }
+
         [Parameter(HelpMessage = "Cmdlet does not return an object by default. If this switch is specified, the cmdlet returns the certificate object that was deleted.")]
         public SwitchParameter PassThru { get; set; }
 
@@ -65,7 +72,25 @@ namespace Microsoft.Azure.Commands.KeyVault
 
         protected override void ProcessRecord()
         {
-            CertificateBundle certBundle = null;
+            if ( InRemovedState.IsPresent )
+            {
+                ConfirmAction(
+                    Force.IsPresent,
+                    string.Format(
+                        CultureInfo.InvariantCulture,
+                        KeyVaultProperties.Resources.RemoveDeletedCertificateWarning,
+                        Name ),
+                    string.Format(
+                        CultureInfo.InvariantCulture,
+                        KeyVaultProperties.Resources.RemoveDeletedCertificateWhatIfMessage,
+                        Name ),
+                    Name,
+                    ( ) => { DataServiceClient.PurgeCertificate( VaultName, Name ); } );
+
+                return;
+            }
+
+            DeletedCertificateBundle certBundle = null;
 
             ConfirmAction(
                 Force.IsPresent,