+++ src/ResourceManager/KeyVault/Commands.KeyVault/Commands/UndoAzureKeyVaultSecretRemoval.cs
@@ -19,6 +19,7 @@ namespace Microsoft.Azure.Commands.KeyVault
 {
     [Cmdlet(VerbsCommon.Undo, "AzureKeyVaultSecretRemoval",
     SupportsShouldProcess = true,
+    ConfirmImpact = ConfirmImpact.Low,
     HelpUri = Constants.KeyVaultHelpUri)]
     [OutputType(typeof(Secret))]
     public class UndoAzureKeyVaultSecretRemoval : KeyVaultCmdletBase