+++ src/ResourceManager/KeyVault/Commands.KeyVault/Models/KeyVaultDataServiceClient.cs
@@ -679,7 +679,7 @@ namespace Microsoft.Azure.Commands.KeyVault.Models
             return certificateBundle;
         }
 
-        public CertificateBundle DeleteCertificate(string vaultName, string certName)
+        public DeletedCertificateBundle DeleteCertificate(string vaultName, string certName)
         {
             if (string.IsNullOrEmpty(vaultName))
                 throw new ArgumentNullException(nameof(vaultName));
@@ -688,7 +688,7 @@ namespace Microsoft.Azure.Commands.KeyVault.Models
 
             string vaultAddress = this.vaultUriHelper.CreateVaultAddress(vaultName);
 
-            CertificateBundle certBundle;
+            DeletedCertificateBundle certBundle;
 
             try
             {
@@ -700,6 +700,25 @@ namespace Microsoft.Azure.Commands.KeyVault.Models
             }
 
             return certBundle;
+        }
+
+        public void PurgeCertificate(string vaultName, string certName)
+        {
+            if ( string.IsNullOrEmpty( vaultName ) )
+                throw new ArgumentNullException( "vaultName" );
+            if ( string.IsNullOrEmpty( certName ) )
+                throw new ArgumentNullException( "certName" );
+
+            string vaultAddress = this.vaultUriHelper.CreateVaultAddress(vaultName);
+
+            try
+            {
+                this.keyVaultClient.PurgeDeletedCertificateAsync( vaultAddress, certName ).GetAwaiter( ).GetResult( );
+            }
+            catch (Exception ex)
+            {
+                throw GetInnerException( ex );
+            }
         }
 
         public CertificateOperation GetCertificateOperation(string vaultName, string certificateName)
@@ -1375,9 +1394,9 @@ namespace Microsoft.Azure.Commands.KeyVault.Models
         public DeletedKeyBundle GetDeletedKey(string vaultName, string keyName)
         {
             if (string.IsNullOrEmpty(vaultName))
-                throw new ArgumentNullException("vaultName");
+                throw new ArgumentNullException(nameof(vaultName));
             if (string.IsNullOrEmpty(keyName))
-                throw new ArgumentNullException("keyName");
+                throw new ArgumentNullException(nameof(keyName));
 
             string vaultAddress = this.vaultUriHelper.CreateVaultAddress(vaultName);
 
@@ -1404,7 +1423,7 @@ namespace Microsoft.Azure.Commands.KeyVault.Models
         public IEnumerable<DeletedKeyIdentityItem> GetDeletedKeys(KeyVaultObjectFilterOptions options)
         {
             if (options == null)
-                throw new ArgumentNullException("options");
+                throw new ArgumentNullException(nameof(options));
 
             if (string.IsNullOrEmpty(options.VaultName))
                 throw new ArgumentException(KeyVaultProperties.Resources.InvalidVaultName);
@@ -1433,9 +1452,9 @@ namespace Microsoft.Azure.Commands.KeyVault.Models
         public DeletedSecret GetDeletedSecret(string vaultName, string secretName)
         {
             if (string.IsNullOrEmpty(vaultName))
-                throw new ArgumentNullException("vaultName");
+                throw new ArgumentNullException(nameof(vaultName));
             if (string.IsNullOrEmpty(secretName))
-                throw new ArgumentNullException("secretName");
+                throw new ArgumentNullException(nameof(secretName));
 
             string vaultAddress = this.vaultUriHelper.CreateVaultAddress(vaultName);
 
@@ -1462,7 +1481,7 @@ namespace Microsoft.Azure.Commands.KeyVault.Models
         public IEnumerable<DeletedSecretIdentityItem> GetDeletedSecrets(KeyVaultObjectFilterOptions options)
         {
             if (options == null)
-                throw new ArgumentNullException("options");
+                throw new ArgumentNullException(nameof(options));
             if (string.IsNullOrEmpty(options.VaultName))
                 throw new ArgumentException(KeyVaultProperties.Resources.InvalidVaultName);
 
@@ -1490,9 +1509,9 @@ namespace Microsoft.Azure.Commands.KeyVault.Models
         public void PurgeKey(string vaultName, string keyName)
         {
             if (string.IsNullOrEmpty(vaultName))
-                throw new ArgumentNullException("vaultName");
+                throw new ArgumentNullException(nameof(vaultName));
             if (string.IsNullOrEmpty(keyName))
-                throw new ArgumentNullException("keyName");
+                throw new ArgumentNullException(nameof(keyName));
 
             string vaultAddress = this.vaultUriHelper.CreateVaultAddress(vaultName);
 
@@ -1509,9 +1528,9 @@ namespace Microsoft.Azure.Commands.KeyVault.Models
         public void PurgeSecret(string vaultName, string secretName)
         {
             if (string.IsNullOrEmpty(vaultName))
-                throw new ArgumentNullException("vaultName");
+                throw new ArgumentNullException(nameof(vaultName));
             if (string.IsNullOrEmpty(secretName))
-                throw new ArgumentNullException("secretName");
+                throw new ArgumentNullException(nameof(secretName));
 
             string vaultAddress = this.vaultUriHelper.CreateVaultAddress(vaultName);
 
@@ -1528,9 +1547,9 @@ namespace Microsoft.Azure.Commands.KeyVault.Models
         public KeyBundle RecoverKey(string vaultName, string keyName)
         {
             if (string.IsNullOrEmpty(vaultName))
-                throw new ArgumentNullException("vaultName");
+                throw new ArgumentNullException(nameof(vaultName));
             if (string.IsNullOrEmpty(keyName))
-                throw new ArgumentNullException("keyName");
+                throw new ArgumentNullException(nameof(keyName));
 
             string vaultAddress = this.vaultUriHelper.CreateVaultAddress(vaultName);
 
@@ -1550,9 +1569,9 @@ namespace Microsoft.Azure.Commands.KeyVault.Models
         public Secret RecoverSecret(string vaultName, string secretName)
         {
             if (string.IsNullOrEmpty(vaultName))
-                throw new ArgumentNullException("vaultName");
+                throw new ArgumentNullException(nameof(vaultName));
             if (string.IsNullOrEmpty(secretName))
-                throw new ArgumentNullException("secretName");
+                throw new ArgumentNullException(nameof(secretName));
 
             string vaultAddress = this.vaultUriHelper.CreateVaultAddress(vaultName);
 
@@ -1567,6 +1586,95 @@ namespace Microsoft.Azure.Commands.KeyVault.Models
             }
 
             return new Secret(recoveredSecret, this.vaultUriHelper);
+        }
+
+        public DeletedCertificateBundle GetDeletedCertificate( string vaultName, string certName )
+        {
+            if ( string.IsNullOrEmpty( vaultName ) )
+                throw new ArgumentNullException( nameof(vaultName) );
+            if ( string.IsNullOrEmpty( certName ) )
+                throw new ArgumentNullException( nameof(certName) );
+
+            string vaultAddress = this.vaultUriHelper.CreateVaultAddress(vaultName);
+
+            DeletedCertificateBundle deletedCertificate;
+            try
+            {
+                deletedCertificate = this.keyVaultClient.GetDeletedCertificateAsync( vaultAddress, certName ).GetAwaiter( ).GetResult( );
+            }
+            catch ( KeyVaultErrorException ex )
+            {
+                if ( ex.Response.StatusCode == HttpStatusCode.NotFound )
+                    return null;
+                else
+                    throw;
+            }
+            catch ( Exception ex )
+            {
+                throw GetInnerException( ex );
+            }
+
+            return deletedCertificate;
+        }
+
+        public IEnumerable<DeletedCertificateIdentityItem> GetDeletedCertificates( KeyVaultObjectFilterOptions options )
+        {
+            /// TODO [dragosav]: the type handling of certificates in the KV PSh client is
+            /// inconsistent with the other entity types, as well as inconsisten within itself.
+            /// For keys and secrets, the wrapper of the data plane client consumes KV models
+            /// and returns PSH models - i.e. type translation occurs in the wrapper. For
+            /// certificates, individual objects are passed through as KV models to the caller
+            /// (which is the cmdlet implementation, and which performs the type translation),
+            /// but collections are translated in the wrapper - i.e. the wrapper returns 
+            /// CertificateBundle (a KV type) but also CertificateIdentityItem (a PSh type).
+            /// Changing this constitutes a breaking change in the IKVDataServiceClient iface,
+            /// which, although unlikely to be used outside of this code, is however public.
+            if ( options == null )
+                throw new ArgumentNullException( nameof( options ) );
+            if ( string.IsNullOrEmpty( options.VaultName ) )
+                throw new ArgumentException( KeyVaultProperties.Resources.InvalidVaultName );
+
+            string vaultAddress = this.vaultUriHelper.CreateVaultAddress(options.VaultName);
+
+            try
+            {
+                IPage<DeletedCertificateItem> result;
+
+                if ( string.IsNullOrEmpty( options.NextLink ) )
+                    result = this.keyVaultClient.GetDeletedCertificatesAsync( vaultAddress ).GetAwaiter( ).GetResult( );
+                else
+                    result = this.keyVaultClient.GetDeletedCertificatesNextAsync( options.NextLink ).GetAwaiter( ).GetResult( );
+
+                options.NextLink = result.NextPageLink;
+                return ( result == null ) ? new List<DeletedCertificateIdentityItem>( ) :
+                    result.Select( ( deletedItem ) => new DeletedCertificateIdentityItem( deletedItem, this.vaultUriHelper ) );
+            }
+            catch ( Exception ex )
+            {
+                throw GetInnerException( ex );
+            }
+        }
+
+        public CertificateBundle RecoverCertificate( string vaultName, string certName )
+        {
+            if ( string.IsNullOrEmpty( vaultName ) )
+                throw new ArgumentNullException( nameof( vaultName ) );
+            if ( string.IsNullOrEmpty( certName ) )
+                throw new ArgumentNullException( nameof( certName ) );
+
+            string vaultAddress = this.vaultUriHelper.CreateVaultAddress(vaultName);
+
+            CertificateBundle recoveredCertificate;
+            try
+            {
+                recoveredCertificate = this.keyVaultClient.RecoverDeletedCertificateAsync( vaultAddress, certName ).GetAwaiter( ).GetResult( );
+            }
+            catch ( Exception ex )
+            {
+                throw GetInnerException( ex );
+            }
+
+            return recoveredCertificate;
         }
 
         private VaultUriHelper vaultUriHelper;